feat: 优惠策略实现

+++ group-buy-market-lpc-infrastructure/src/main/java/top/licodetech/market/infrastructure/dao/po/GroupBuyDiscount.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import top.licodetech.market.domain.activity.model.valobj.DiscountTypeEnum;
 
 import java.util.Date;
 
@@ -23,7 +24,7 @@ public class GroupBuyDiscount {
     /** 折扣描述 */
     private String discountDesc;
     /** 折扣类型（0:base、1:tag） */
-    private Byte discountType;
+    private Integer discountType;
     /** 营销优惠计划（ZJ:直减、MJ:满减、N元购） */
     private String marketPlan;
     /** 营销优惠表达式 */
